Iterface folders has been renamed to Contracts
