+++ CarRental.DLL/Contracts/IGenericRepository.cs
@@ -1,6 +1,6 @@
 ﻿using CarRental.DLL.Entities;
 
-namespace CarRental.DLL.Interfaces
+namespace CarRental.DLL.Contracts
 {
     public interface IGenericRepository<T> where T : BaseEntity
     {
