+++ CarRental.DLL/Contracts/IManufacturerRepository.cs
@@ -1,6 +1,6 @@
 ﻿using Car_Rental.DLL.Entities;
 
-namespace CarRental.DLL.Interfaces
+namespace CarRental.DLL.Contracts
 {
     public interface IManufacturerRepository : IGenericRepository<Manufacturer>
     {
