+++ CarRental.DLL/Contracts/IUnitOfWork.cs
@@ -1,4 +1,4 @@
-﻿namespace CarRental.DLL.Interfaces
+﻿namespace CarRental.DLL.Contracts
 {
     public interface IUnitOfWork : IDisposable
     {
