+++ CarRental/Controllers/ManufacturerController.cs
@@ -1,15 +1,15 @@
-﻿using CarRental.BLL.Interfaces;
+﻿using CarRental.BLL.Contracts;
 using Microsoft.AspNetCore.Mvc;
 
 namespace CarRental.Controllers
 {
     [Route("api/[controller]")]
     [ApiController]
-    public class ManufacturerController : ControllerBase
+    public class ManufacturersController : ControllerBase
     {
         private readonly IManufacturerService _manufacturerService;
 
-        public ManufacturerController(IManufacturerService manufacturerService)
+        public ManufacturersController(IManufacturerService manufacturerService)
         {
             _manufacturerService = manufacturerService;
         }
